+++ frontend/src/pages/dreams/mine/MyDreamPage.tsx
@@ -1,9 +1,11 @@
 import MyDreamCards from "./MyDreamCards";
 import MyDreamInput from "./MyDreamInput";
+import Header from "../components/header";
 
 const MyDreamPage = () => {
   return (
     <div>
+      <Header/>
       <MyDreamInput />
       <MyDreamCards />
     </div>
